+++ CHANGELOG.md
@@ -8,4 +8,54 @@
 
 #### New
 
+- Renamed the available castbar particle themes to their new player-facing names
+- Removed unused castbar particle themes from the selector
+
 #### Other
+
+- Castbar particle theme renames:
+Aim -> Hunt
+Alliance -> Blue Sparkle
+Arcane -> Arcane
+Arctic -> Frost
+Bronze -> Gold Sparkle
+Chaos -> Green Gas
+Chi'Ji -> Soft Leaves
+Earth -> Earth
+Felfire -> Green Fire
+Fire -> Fire
+Fishing -> Water
+Fist of Fury -> Green & Blue Leaves
+Frostfire -> Purple Glow
+Herbalism -> Leaves
+Holy -> Holy
+Horde -> Red Sparkle
+Lumber -> Wood
+Mining -> Rock
+Mistweaver -> Green Mist
+Moon -> Lunar
+Mossy Stone -> Celestial Green
+Neutral -> Sparkle
+Sacred -> Holy 2
+Shadow -> Shadow
+Thunder -> Blue Sparkle
+Void -> Void
+
+- Removed castbar particle themes:
+Arcaneum
+Engrenages
+Frost
+Honey
+Inferno
+Lava
+Metal Icon
+Nature
+Neutral Metal
+Neutral 2
+Neutral 3
+Skinning
+Viking Icon
+
+
+
+